+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.6.0 - 2021-1-5
+### Added
+- Added support for custom csr yaml
+- Added new nodes in isa_schema: pmp_granularity and physical_addr_sz
+- Checks for pmp, counters and custom csrs
+- medeleg, mideleg check for S or N extension
+- updated the warl syntax slightly for easier parsing.
+
+### Changed
+- fixed warl parsing and islegal function to check reset values 
+
 ## 2.5.1 - 2020-11-6
 ### Changed
 - modified sn_check and su_check